app/backend/requirements.txt
# Core Framework
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# AI & LangChain
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-groq>=0.0.1
openai>=1.3.0

# Realtime Voice
websockets>=12.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
aiosqlite>=0.19.0

# External APIs  
groq>=0.4.0
tavily-python>=0.3.0
together>=0.2.0

# Async & Utilities
asyncio-mqtt>=0.13.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Monitoring
structlog>=23.2.0